+++ SCMM.Web/Server/Domain/Models/Steam/SteamStoreItem.cs
@@ -8,6 +8,6 @@ namespace SCMM.Web.Server.Domain.Models.Steam
 
         public SteamCurrency Currency { get; set; }
 
-        public int StorePrice { get; set; }
+        public long StorePrice { get; set; }
     }
 }
