+++ SCMM.Web/Shared/Domain/DTOs/StoreItems/StoreItemListDTO.cs
@@ -21,7 +21,7 @@ namespace SCMM.Web.Shared.Domain.DTOs.StoreItems
 
         public CurrencyDTO Currency { get; set; }
 
-        public int StorePrice { get; set; }
+        public long StorePrice { get; set; }
 
         public int MarketRankPosition { get; set; }
